requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
schedule==1.2.2
python-dotenv==1.0.1
python-docx>=1.1.0
